% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_hostname.R
\name{matomo_hostname}
\alias{matomo_hostname}
\title{Matomo hostname}
\usage{
matomo_hostname()
}
\value{
A character string containing the Matomo hostname.
}
\description{
This function returns the Matomo hostname that is set in the environment variable MATOMO_HOST.
}
\examples{
Sys.setenv(MATOMO_HOST = "https://demo.matomo.org")

matomo_hostname()
}
