% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_media_video_titles.R
\name{matomo_media_video_titles}
\alias{matomo_media_video_titles}
\title{Query the Matomo Media Analytics for video resources}
\usage{
matomo_media_video_titles(
  idSite = NULL,
  period = "day",
  date = "yesterday",
  segment = NULL,
  idSubtable = NULL,
  secondaryDimension = NULL,
  ...
)
}
\arguments{
\item{idSite}{The id of your website.}

\item{period}{The period you request the statistics for. Can be any of: day, week, month, year or range.}

\item{date}{The date you request the statictics for. Standard format = YYYY-MM-DD or today or yesterday.}

\item{segment}{Defines the Custom Segment you wish to filter your reports to.}

\item{idSubtable}{The id of a possible subtable.}

\item{secondaryDimension}{Secondary dimension is the dimension used in the sub-table of the Event report you are requesting.}

\item{...}{Possible further options to the query in the format variable = value. For a list of possible options see: \url{https://developer.matomo.org/api-reference/reporting-api}}
}
\value{
A list containing a data frame with the requested analytics values.
}
\description{
Returns the media analytics for video resources.
}
\details{
This function expects the hostname and token to be set as environment variables.
}
\examples{
Sys.setenv(MATOMO_HOST = "https://demo.matomo.org")
Sys.setenv(MATOMO_TOKEN = "anonymous")

matomo_media_video_titles(idSite = 62)
}
