% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_package_query.R
\name{matomo_package_query}
\alias{matomo_package_query}
\title{matomoR query function}
\usage{
matomo_package_query(url = NULL, query = NULL)
}
\arguments{
\item{url}{The URL for the api call.}

\item{query}{A list of query parameters.}
}
\value{
The response of the api call.
}
\description{
This is an internal function of the matomoR package to query the api.
}
\examples{
Sys.setenv(MATOMO_HOST = "https://demo.matomo.org")
Sys.setenv(MATOMO_TOKEN = "anonymous")

# Construct the query
query <- list(
 module = "API",
 method = "API.getMatomoVersion"
)

# Create the url
url <- httr::modify_url(matomo_hostname(), query = query)

# Make the query
matomo_package_query(url)
}
