% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_package_useragent.R
\name{matomo_package_useragent}
\alias{matomo_package_useragent}
\title{matomoR user agent}
\usage{
matomo_package_useragent()
}
\value{
A character string containing the user agent used in api calls generate with the matomoR package.
}
\description{
This function returns the user agent used in api calls of the matomoR package.
}
\examples{
matomo_package_useragent()
}
