% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_referrers.R
\name{matomo_referrers_all}
\alias{matomo_referrers_all}
\title{Query the Matomo Referrer Analytics}
\usage{
matomo_referrers_all(
  idSite = NULL,
  period = "day",
  date = "yesterday",
  segment = NULL
)
}
\arguments{
\item{idSite}{The id of your website.}

\item{period}{The period you request the statistics for. Can be any of: day, week, month, year or range.}

\item{date}{The date you request the statictics for. Standard format = YYYY-MM-DD or today or yesterday.}

\item{segment}{Defines the Custom Segment you wish to filter your reports to.}
}
\value{

}
\description{
Returns all referrers data
}
