% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_token.R
\name{matomo_token}
\alias{matomo_token}
\title{Matomo Token}
\usage{
matomo_token()
}
\value{
A character string containing the Matomo token.
}
\description{
This function returns the Matomo access token that is set in the environment variable MATOMO_TOKEN.
}
\examples{
Sys.setenv(MATOMO_TOKEN = "anonymous")

matomo_token()
}
