% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_version.R
\name{matomo_version}
\alias{matomo_version}
\title{Query Matomo for the version}
\usage{
matomo_version()
}
\value{
A list containing the server version.
}
\description{
Returns the version of the Matomo server.
}
\details{
This function expects the hostname and token to be set as environment variables.
}
\examples{
Sys.setenv(MATOMO_HOST = "https://demo.matomo.org")
Sys.setenv(MATOMO_TOKEN = "anonymous")

matomo_version()
}
