% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomoR-package.R
\docType{package}
\name{matomoR-package}
\alias{matomoR}
\alias{matomoR-package}
\title{matomoR: 'Matomo' Analytics Web API Package}
\description{
Query the Analyics API of a 'Matomo' server. For more information about 'Matomo' and the api see: <https://matomo.org/docs/analytics-api/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ebbertd/matomoR}
  \item Report bugs at \url{https://github.com/ebbertd/matomoR/issues}
}

}
\author{
\strong{Maintainer}: Daniel Ebbert \email{daniel.ebbert@uni-muenster.de} (\href{https://orcid.org/0000-0003-3666-7205}{ORCID})

}
\keyword{internal}
