% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_media_video_resources.R
\name{print.matomo_media_video_resources}
\alias{print.matomo_media_video_resources}
\title{Print result of matomo_media_video_resources()}
\usage{
\method{print}{matomo_media_video_resources}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the matomo_media_video_resources() function.
}
\description{
Print a structured return of the matomo_media_video_resources() function.
}
\examples{
Sys.setenv(MATOMO_HOST = "https://demo.matomo.org")
Sys.setenv(MATOMO_TOKEN = "anonymous")

resp <- matomo_media_video_resources(idSite = 62)

resp
}
\seealso{
\code{\link{matomo_media_video_resources}}
}
