% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_version.R
\name{print.matomo_version}
\alias{print.matomo_version}
\title{Print result of matomo_version()}
\usage{
\method{print}{matomo_version}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the matomo_version() function.
}
\description{
Print a structured return of the matomo_version() function.
}
\examples{
Sys.setenv(MATOMO_HOST = "https://demo.matomo.org")
Sys.setenv(MATOMO_TOKEN = "anonymous")

resp <- matomo_version()

resp
}
\seealso{
\code{\link{matomo_version}}
}
