% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo_package_query.R
\name{matomo_package_query}
\alias{matomo_package_query}
\title{matomoR query function}
\usage{
matomo_package_query(query = NULL, url = NULL)
}
\arguments{
\item{query}{A list of query parameters.}

\item{url}{The URL for the api call, by default the environment variable \code{MATOMO_HOST}.}
}
\value{
The response of the api call.
}
\description{
This is an internal function of the matomoR package to query the api.
You can use it to call any endpoint that does not have a higher-level interface.
}
\examples{
Sys.setenv(MATOMO_HOST = "https://demo.matomo.org")
Sys.setenv(MATOMO_TOKEN = "anonymous")

# Construct the query
query <- list(
 module = "API",
 method = "API.getMatomoVersion"
)

# Make the query
matomo_package_query(query = query)
}
